HuffmanTableLUT: pacify harmless implicit conversion sanitizer diag
../src/librawspeed/decompressors/HuffmanTableLUT.h:140:22: runtime error: implicit conversion from type 'int' of value -32768 (32-bit, signed) to type 'uint32_t' (aka 'unsigned int') changed the value to 4294934528 (32-bit, unsigned)

+++ src/librawspeed/decompressors/HuffmanTableLUT.h
@@ -137,7 +137,7 @@ public:
                                      /*effectiveBitwidth=*/LookupDepth);
               diff &= ((1 << diff_l) - 1);
             } else
-              diff = -32768;
+              diff = uint32_t(-32768);
             decodeLookup[c] |= static_cast<int32_t>(
                 static_cast<uint32_t>(extend(diff, diff_l)) << PayloadShift);
           }
